+++ R/utils.R
@@ -328,7 +328,8 @@ h_confirm_large_levels <- function(x) {
 }
 
 #' Default Value on NULL
-#' @description Return default value when first argument is NULL.
+#' @description `r lifecycle::badge("experimental")`
+#' Return default value when first argument is NULL.
 #'
 #' @param x Object.
 #' @param y Object.